# Fixes #1113- Adding multi-id qualification to stored instances in lifetime scope