--- a/src/Autofac/Core/ISharingLifetimeScope.cs
+++ b/src/Autofac/Core/ISharingLifetimeScope.cs
@@ -51,11 +51,31 @@ namespace Autofac.Core
         bool TryGetSharedInstance(Guid id, out object value);
 
         /// <summary>
+        /// Try to retrieve a shared instance based on a primary GUID key and
+        /// possible secondary qualifying GUID key.
+        /// </summary>
+        /// <param name="primaryId">Key to look up.</param>
+        /// <param name="qualifyingId">Secondary key to look up.</param>
+        /// <param name="value">The instance that has the specified key.</param>
+        /// <returns><c>true</c> if the key was found; otherwise, <c>false</c>.</returns>
+        bool TryGetSharedInstance(Guid primaryId, Guid? qualifyingId, out object value);
+
+        /// <summary>
         /// Creates a shared instance with a GUID key.
         /// </summary>
         /// <param name="id">Key.</param>
         /// <param name="creator">A function that will create the instance when called.</param>
         /// <returns>The shared instance.</returns>
         object CreateSharedInstance(Guid id, Func<object> creator);
+
+        /// <summary>
+        /// Creates a shared instance with a primary GUID key and
+        /// possible secondary qualifying GUID key.
+        /// </summary>
+        /// <param name="primaryId">Key.</param>
+        /// <param name="qualifyingId">Secondary key to look up.</param>
+        /// <param name="creator">A function that will create the instance when called.</param>
+        /// <returns>The shared instance.</returns>
+        object CreateSharedInstance(Guid primaryId, Guid? qualifyingId, Func<object> creator);
     }
 }