--- a/src/Autofac/Core/Lifetime/LifetimeScope.cs
+++ b/src/Autofac/Core/Lifetime/LifetimeScope.cs
@@ -48,7 +48,8 @@ namespace Autofac.Core.Lifetime
         /// Protects shared instances from concurrent access. Other members and the base class are threadsafe.
         /// </summary>
         private readonly object _synchRoot = new object();
-        private readonly ConcurrentDictionary<(Guid, Guid?), object> _sharedInstances = new ConcurrentDictionary<(Guid, Guid?), object>();
+        private readonly ConcurrentDictionary<Guid, object> _sharedInstances = new ConcurrentDictionary<Guid, object>();
+        private readonly ConcurrentDictionary<(Guid, Guid), object> _sharedQualifiedInstances = new ConcurrentDictionary<(Guid, Guid), object>();
         private object? _anonymousTag;
         private LifetimeScope? parentScope;
 
@@ -82,7 +83,7 @@ namespace Autofac.Core.Lifetime
         /// <param name="componentRegistry">Components used in the scope.</param>
         public LifetimeScope(IComponentRegistry componentRegistry, object tag)
         {
-            _sharedInstances[(SelfRegistrationId, null)] = this;
+            _sharedInstances[SelfRegistrationId] = this;
             ComponentRegistry = componentRegistry ?? throw new ArgumentNullException(nameof(componentRegistry));
             Tag = tag ?? throw new ArgumentNullException(nameof(tag));
             RootLifetimeScope = this;
@@ -292,35 +293,56 @@ namespace Autofac.Core.Lifetime
         /// <inheritdoc />
         public object CreateSharedInstance(Guid id, Func<object> creator)
         {
-            return CreateSharedInstance(id, null, creator);
+            if (creator == null) throw new ArgumentNullException(nameof(creator));
+            lock (_synchRoot)
+            {
+                if (_sharedInstances.TryGetValue(id, out var result)) return result;
+
+                result = creator();
+                if (_sharedInstances.ContainsKey(id))
+                    throw new DependencyResolutionException(string.Format(CultureInfo.CurrentCulture, LifetimeScopeResources.SelfConstructingDependencyDetected, result.GetType().FullName));
+
+                _sharedInstances.TryAdd(id, result);
+
+                return result;
+            }
         }
 
         /// <inheritdoc/>
         public object CreateSharedInstance(Guid primaryId, Guid? qualifyingId, Func<object> creator)
         {
             if (creator == null) throw new ArgumentNullException(nameof(creator));
+            if (qualifyingId == null)
+            {
+                return CreateSharedInstance(primaryId, creator);
+            }
 
             lock (_synchRoot)
             {
-                var instanceKey = (primaryId, qualifyingId);
+                var instanceKey = (primaryId, qualifyingId.Value);
 
-                if (_sharedInstances.TryGetValue(instanceKey, out var result)) return result;
+                if (_sharedQualifiedInstances.TryGetValue(instanceKey, out var result)) return result;
 
                 result = creator();
-                if (_sharedInstances.ContainsKey(instanceKey))
+                if (_sharedQualifiedInstances.ContainsKey(instanceKey))
                     throw new DependencyResolutionException(string.Format(CultureInfo.CurrentCulture, LifetimeScopeResources.SelfConstructingDependencyDetected, result.GetType().FullName));
 
-                _sharedInstances.TryAdd(instanceKey, result);
+                _sharedQualifiedInstances.TryAdd(instanceKey, result);
 
                 return result;
             }
         }
 
         /// <inheritdoc />
-        public bool TryGetSharedInstance(Guid id, out object value) => TryGetSharedInstance(id, null, out value);
+        public bool TryGetSharedInstance(Guid id, out object value) => _sharedInstances.TryGetValue(id, out value);
 
         /// <inheritdoc/>
-        public bool TryGetSharedInstance(Guid primaryId, Guid? qualifyingId, out object value) => _sharedInstances.TryGetValue((primaryId, qualifyingId), out value);
+        public bool TryGetSharedInstance(Guid primaryId, Guid? qualifyingId, out object value)
+        {
+            return qualifyingId == null
+                ? TryGetSharedInstance(primaryId, out value)
+                : _sharedQualifiedInstances.TryGetValue((primaryId, qualifyingId.Value), out value);
+        }
 
         /// <summary>
         /// Gets the disposer associated with this container. Instances can be associated