Fixes #1113
As noted in the investigation of #1113, the fact that shared instances are only stored with the registration ID results in some bad behavior when multiple services end up being decorated via Resolve<IEnumerable<T>> and some of the targets are shared.

After trying a fix in #1114, the feedback pointed out the original approach was too expensive for such a commonly executed path. This PR uses the approach that was suggested in the original PR with the below benchmarks. It seems that this change does cost an additional ~300ns per lookup/resolution performed.

# Original
<details><summary>Expand</summary>

```Autofac.Benchmarks.ChildScopeResolveBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.201
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|  Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
| Resolve | 41.08 us | 0.203 us | 0.180 us | 6.1646 | 0.3662 |     - |   50.7 KB |

```Autofac.Benchmarks.ConcurrencyBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.201
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                          Method | ResolveTaskCount | ResolvesPerTask |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------------------------- |----------------- |---------------- |---------:|----------:|----------:|---------:|--------:|------:|----------:|
| MultipleResolvesOnMultipleTasks |              100 |             100 | 2.041 ms | 0.0402 ms | 0.0376 ms | 617.1875 | 27.3438 |     - |    4.9 MB |

```Autofac.Benchmarks.Decorators.KeylessNestedSharedInstanceLambdaBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **477.0 ns** |  **2.84 ns** |  **2.52 ns** |  **1.00** |    **0.00** | **0.2093** |      **-** |     **-** |   **1.71 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **4,066.8 ns** | **43.22 ns** | **40.43 ns** |     **?** |       **?** | **0.9613** | **0.0076** |     **-** |    **7.9 KB** |
|           ResolveT |           1 | 2,151.8 ns |  9.01 ns |  8.43 ns |     ? |       ? | 0.5875 | 0.0038 |     - |   4.83 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **6,679.9 ns** | **37.64 ns** | **35.21 ns** |     **?** |       **?** | **1.5945** | **0.0153** |     **-** |  **13.05 KB** |
|           ResolveT |           2 | 3,270.0 ns | 12.28 ns | 11.49 ns |     ? |       ? | 0.8812 | 0.0076 |     - |    7.2 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **9,114.3 ns** | **39.82 ns** | **37.25 ns** |     **?** |       **?** | **2.2278** | **0.0153** |     **-** |  **18.21 KB** |
|           ResolveT |           3 | 4,120.2 ns |  7.43 ns |  6.95 ns |     ? |       ? | 1.1673 | 0.0076 |     - |   9.58 KB |

```Autofac.Benchmarks.PropertyInjectionBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|  Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |---------:|----------:|----------:|-------:|-------:|------:|----------:|
| Resolve | 3.635 us | 0.0158 us | 0.0148 us | 0.7706 | 0.0038 |     - |    6.3 KB |

```Autofac.Benchmarks.OpenGenericBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ResolveOpenGeneric | 328.5 ns | 5.94 ns | 5.56 ns | 0.0820 |     - |     - |     688 B |

```Autofac.Benchmarks.DeepGraphResolveBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|  Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
| Resolve | 11.31 us | 0.036 us | 0.034 us | 2.2430 | 0.0458 |     - |  18.45 KB |

```Autofac.Benchmarks.ConcurrencyBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                          Method | ResolveTaskCount | ResolvesPerTask |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------------------------- |----------------- |---------------- |---------:|----------:|----------:|---------:|--------:|------:|----------:|
| MultipleResolvesOnMultipleTasks |              100 |             100 | 1.941 ms | 0.0386 ms | 0.0460 ms | 617.1875 | 27.3438 |     - |    4.9 MB |

```Autofac.Benchmarks.Decorators.KeylessSimpleLambdaBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **479.6 ns** |  **2.39 ns** |  **2.24 ns** |  **1.00** |    **0.00** | **0.2089** |      **-** |     **-** |   **1.71 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **2,985.5 ns** | **22.68 ns** | **21.21 ns** |     **?** |       **?** | **0.6714** | **0.0038** |     **-** |   **5.49 KB** |
|           ResolveT |           1 | 1,399.8 ns |  8.95 ns |  8.37 ns |     ? |       ? | 0.4158 |      - |     - |    3.4 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **5,271.4 ns** | **30.99 ns** | **28.99 ns** |     **?** |       **?** | **1.1292** | **0.0076** |     **-** |   **9.27 KB** |
|           ResolveT |           2 | 2,263.0 ns | 12.28 ns | 11.48 ns |     ? |       ? | 0.6218 | 0.0038 |     - |   5.09 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **7,591.6 ns** | **25.30 ns** | **23.66 ns** |     **?** |       **?** | **1.5945** | **0.0153** |     **-** |  **13.05 KB** |
|           ResolveT |           3 | 3,122.5 ns | 18.08 ns | 16.03 ns |     ? |       ? | 0.8278 | 0.0038 |     - |   6.77 KB |

```Autofac.Benchmarks.Decorators.KeylessNestedLambdaBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |        Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |------------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |    **478.3 ns** |  **1.47 ns** |  **1.37 ns** |  **1.00** |    **0.00** | **0.2089** |      **-** |     **-** |   **1.71 KB** |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** |  **4,070.9 ns** |  **9.48 ns** |  **8.87 ns** |     **?** |       **?** | **0.9308** | **0.0076** |     **-** |   **7.65 KB** |
|           ResolveT |           1 |  1,894.9 ns |  9.34 ns |  8.74 ns |     ? |       ? | 0.5474 | 0.0038 |     - |   4.48 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** |  **7,164.8 ns** | **18.07 ns** | **16.90 ns** |     **?** |       **?** | **1.6632** | **0.0153** |     **-** |  **13.59 KB** |
|           ResolveT |           2 |  3,293.6 ns | 14.04 ns | 13.13 ns |     ? |       ? | 0.8850 | 0.0076 |     - |   7.24 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **10,575.3 ns** | **23.66 ns** | **22.14 ns** |     **?** |       **?** | **2.3804** | **0.0153** |     **-** |  **19.52 KB** |
|           ResolveT |           3 |  4,532.6 ns | 27.35 ns | 25.58 ns |     ? |       ? | 1.2207 | 0.0076 |     - |  10.01 KB |

```Autofac.Benchmarks.RootContainerResolveBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                     Method |       Mean |     Error |    StdDev |  Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |-----------:|----------:|----------:|-------:|--------:|-------:|------:|------:|----------:|
|                OperatorNew |   2.571 ns | 0.0302 ns | 0.0283 ns |   1.00 |    0.00 | 0.0029 |     - |     - |      24 B |
| NonSharedReflectionResolve | 335.762 ns | 2.2634 ns | 2.1172 ns | 130.62 |    1.70 | 0.0820 |     - |     - |     688 B |
|   NonSharedDelegateResolve | 253.400 ns | 1.2290 ns | 1.1496 ns |  98.58 |    1.05 | 0.0639 |     - |     - |     536 B |
|              SharedResolve | 229.811 ns | 0.5025 ns | 0.4700 ns |  89.40 |    0.95 | 0.0610 |     - |     - |     512 B |

```Autofac.Benchmarks.Decorators.KeylessNestedSharedInstanceBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **477.8 ns** |  **2.03 ns** |  **1.80 ns** |  **1.00** |    **0.00** | **0.2089** |      **-** |     **-** |   **1.71 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **4,340.2 ns** | **15.22 ns** | **14.24 ns** |     **?** |       **?** | **1.0376** | **0.0076** |     **-** |   **8.48 KB** |
|           ResolveT |           1 | 2,268.8 ns |  5.55 ns |  5.19 ns |     ? |       ? | 0.6599 | 0.0038 |     - |   5.41 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **6,719.7 ns** | **28.59 ns** | **26.75 ns** |     **?** |       **?** | **1.6632** | **0.0153** |     **-** |  **13.63 KB** |
|           ResolveT |           2 | 3,247.4 ns |  7.98 ns |  7.46 ns |     ? |       ? | 0.9499 | 0.0076 |     - |   7.78 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **8,997.0 ns** | **37.94 ns** | **35.49 ns** |     **?** |       **?** | **2.2888** | **0.0153** |     **-** |  **18.79 KB** |
|           ResolveT |           3 | 4,280.1 ns | 15.04 ns | 14.07 ns |     ? |       ? | 1.2360 | 0.0076 |     - |  10.16 KB |

```Autofac.Benchmarks.ConcurrencyNestedScopeBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                          Method | ConcurrentRequests | RepeatCount |     Mean |     Error |    StdDev |   Median |    Gen 0 |    Gen 1 | Gen 2 | Allocated |
|-------------------------------- |------------------- |------------ |---------:|----------:|----------:|---------:|---------:|---------:|------:|----------:|
| MultipleResolvesOnMultipleTasks |                100 |          10 | 1.820 ms | 0.0358 ms | 0.0536 ms | 1.793 ms | 863.2813 | 156.2500 |     - |   6.83 MB |

```Autofac.Benchmarks.Decorators.KeyedNestedBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **476.0 ns** |  **1.29 ns** |  **1.20 ns** |  **1.00** |    **0.00** | **0.2093** | **0.0005** |     **-** |   **1.71 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **2,772.6 ns** |  **8.81 ns** |  **8.24 ns** |     **?** |       **?** | **0.5455** | **0.0038** |     **-** |   **4.48 KB** |
|           ResolveT |           1 | 1,345.6 ns |  4.43 ns |  4.14 ns |     ? |       ? | 0.3490 | 0.0019 |     - |   2.85 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **4,917.2 ns** | **14.56 ns** | **13.62 ns** |     **?** |       **?** | **0.8850** | **0.0076** |     **-** |   **7.26 KB** |
|           ResolveT |           2 | 2,101.8 ns | 13.48 ns | 12.61 ns |     ? |       ? | 0.4883 |      - |     - |   3.99 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **6,977.8 ns** | **25.24 ns** | **23.61 ns** |     **?** |       **?** | **1.2207** | **0.0076** |     **-** |  **10.03 KB** |
|           ResolveT |           3 | 2,929.2 ns |  9.47 ns |  8.86 ns |     ? |       ? | 0.6256 | 0.0038 |     - |   5.13 KB |

```Autofac.Benchmarks.Decorators.KeylessNestedBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |        Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |------------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |    **479.3 ns** |  **1.24 ns** |  **1.16 ns** |  **1.00** |    **0.00** | **0.2089** |      **-** |     **-** |   **1.71 KB** |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** |  **4,198.6 ns** | **11.59 ns** | **10.84 ns** |     **?** |       **?** | **1.0757** | **0.0076** |     **-** |    **8.8 KB** |
|           ResolveT |           1 |  1,998.2 ns |  6.73 ns |  6.30 ns |     ? |       ? | 0.6180 | 0.0038 |     - |   5.05 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** |  **7,809.8 ns** | **21.39 ns** | **20.01 ns** |     **?** |       **?** | **1.9379** | **0.0153** |     **-** |   **15.9 KB** |
|           ResolveT |           2 |  3,404.2 ns | 10.92 ns | 10.22 ns |     ? |       ? | 1.0262 | 0.0076 |     - |    8.4 KB |
…