@VonOgre 2020-05-10
Benchmarks after Feedback:
<details><summary>Expand</summary>

```Autofac.Benchmarks.ChildScopeResolveBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|  Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
| Resolve | 40.65 us | 0.190 us | 0.177 us | 6.4087 | 0.4272 |     - |  52.63 KB |

```Autofac.Benchmarks.ConcurrencyBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                          Method | ResolveTaskCount | ResolvesPerTask |     Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------------------------- |----------------- |---------------- |---------:|----------:|----------:|---------:|--------:|------:|----------:|
| MultipleResolvesOnMultipleTasks |              100 |             100 | 1.961 ms | 0.0373 ms | 0.0414 ms | 617.1875 | 27.3438 |     - |    4.9 MB |

```Autofac.Benchmarks.ConcurrencyNestedScopeBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                          Method | ConcurrentRequests | RepeatCount |     Mean |     Error |    StdDev |   Median |     Gen 0 |    Gen 1 | Gen 2 | Allocated |
|-------------------------------- |------------------- |------------ |---------:|----------:|----------:|---------:|----------:|---------:|------:|----------:|
| MultipleResolvesOnMultipleTasks |                100 |          10 | 1.899 ms | 0.0386 ms | 0.0903 ms | 1.848 ms | 1101.5625 | 250.0000 |     - |   8.71 MB |

```Autofac.Benchmarks.Decorators.KeyedGenericBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **591.4 ns** |  **2.17 ns** |  **2.03 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **2,782.1 ns** | **10.19 ns** |  **9.53 ns** |     **?** |       **?** | **0.6866** | **0.0076** |     **-** |   **5.63 KB** |
|           ResolveT |           1 | 1,345.6 ns |  5.43 ns |  5.08 ns |     ? |       ? | 0.4768 | 0.0038 |     - |   3.91 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **4,813.4 ns** | **45.97 ns** | **43.00 ns** |     **?** |       **?** | **1.0452** | **0.0076** |     **-** |   **8.59 KB** |
|           ResolveT |           2 | 2,044.4 ns |  7.06 ns |  6.60 ns |     ? |       ? | 0.6256 | 0.0038 |     - |   5.14 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **6,663.6 ns** | **38.90 ns** | **36.39 ns** |     **?** |       **?** | **1.4114** | **0.0153** |     **-** |  **11.55 KB** |
|           ResolveT |           3 | 2,736.6 ns | 10.91 ns | 10.20 ns |     ? |       ? | 0.7782 | 0.0076 |     - |   6.38 KB |

```Autofac.Benchmarks.Decorators.KeyedNestedBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **595.4 ns** |  **1.71 ns** |  **1.52 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **2,965.5 ns** | **18.06 ns** | **16.89 ns** |     **?** |       **?** | **0.6638** | **0.0076** |     **-** |   **5.45 KB** |
|           ResolveT |           1 | 1,476.8 ns |  5.32 ns |  4.97 ns |     ? |       ? | 0.4654 | 0.0038 |     - |   3.81 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **5,223.9 ns** | **12.67 ns** | **11.85 ns** |     **?** |       **?** | **0.9995** | **0.0076** |     **-** |   **8.22 KB** |
|           ResolveT |           2 | 2,268.8 ns |  6.80 ns |  6.36 ns |     ? |       ? | 0.6027 | 0.0038 |     - |   4.95 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **7,245.8 ns** |  **9.09 ns** |  **8.50 ns** |     **?** |       **?** | **1.3428** | **0.0153** |     **-** |  **10.99 KB** |
|           ResolveT |           3 | 3,084.1 ns |  8.55 ns |  7.99 ns |     ? |       ? | 0.7439 | 0.0076 |     - |   6.09 KB |

```Autofac.Benchmarks.Decorators.KeyedSimpleBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **588.5 ns** |  **1.78 ns** |  **1.67 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **2,422.1 ns** | **10.17 ns** |  **9.51 ns** |     **?** |       **?** | **0.6065** | **0.0038** |     **-** |   **4.98 KB** |
|           ResolveT |           1 | 1,202.8 ns |  3.43 ns |  3.21 ns |     ? |       ? | 0.4368 | 0.0038 |     - |   3.58 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **4,194.2 ns** | **19.81 ns** | **18.53 ns** |     **?** |       **?** | **0.8850** | **0.0076** |     **-** |   **7.28 KB** |
|           ResolveT |           2 | 1,777.1 ns |  4.19 ns |  3.92 ns |     ? |       ? | 0.5474 | 0.0057 |     - |   4.48 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **5,887.4 ns** | **23.91 ns** | **22.36 ns** |     **?** |       **?** | **1.1673** | **0.0076** |     **-** |   **9.59 KB** |
|           ResolveT |           3 | 2,313.9 ns | 13.31 ns | 12.45 ns |     ? |       ? | 0.6599 | 0.0038 |     - |   5.39 KB |

```Autofac.Benchmarks.Decorators.KeylessGenericBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |-----------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |   **601.4 ns** |  **2.78 ns** |  **2.60 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** | **3,008.0 ns** | **13.24 ns** | **11.74 ns** |     **?** |       **?** | **0.8240** | **0.0076** |     **-** |   **6.75 KB** |
|           ResolveT |           1 | 1,504.8 ns |  4.48 ns |  3.97 ns |     ? |       ? | 0.5512 | 0.0038 |     - |   4.51 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** | **5,149.6 ns** | **21.39 ns** | **20.01 ns** |     **?** |       **?** | **1.3199** | **0.0153** |     **-** |  **10.83 KB** |
|           ResolveT |           2 | 2,351.5 ns |  6.18 ns |  5.48 ns |     ? |       ? | 0.7744 | 0.0076 |     - |   6.34 KB |
|                    |             |            |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **7,664.0 ns** | **24.86 ns** | **23.25 ns** |     **?** |       **?** | **1.8158** | **0.0153** |     **-** |  **14.91 KB** |
|           ResolveT |           3 | 3,187.1 ns | 13.51 ns | 12.64 ns |     ? |       ? | 0.9995 | 0.0114 |     - |   8.18 KB |

```Autofac.Benchmarks.Decorators.KeylessNestedBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |        Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |------------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |    **600.6 ns** |  **1.36 ns** |  **1.27 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** |  **4,510.4 ns** | **15.43 ns** | **14.43 ns** |     **?** |       **?** | **1.1902** | **0.0153** |     **-** |   **9.77 KB** |
|           ResolveT |           1 |  2,253.1 ns | 12.53 ns | 10.46 ns |     ? |       ? | 0.7362 | 0.0076 |     - |   6.02 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** |  **8,139.7 ns** | **33.82 ns** | **31.64 ns** |     **?** |       **?** | **2.0599** | **0.0153** |     **-** |  **16.86 KB** |
|           ResolveT |           2 |  3,697.5 ns | 12.41 ns | 11.60 ns |     ? |       ? | 1.1444 | 0.0114 |     - |   9.36 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **11,702.5 ns** | **53.38 ns** | **49.93 ns** |     **?** |       **?** | **2.9297** | **0.0153** |     **-** |  **23.95 KB** |
|           ResolveT |           3 |  5,175.0 ns | 32.95 ns | 30.82 ns |     ? |       ? | 1.5488 | 0.0076 |     - |   12.7 KB |

```Autofac.Benchmarks.Decorators.KeylessNestedLambdaBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |        Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |------------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |    **630.1 ns** |  **2.29 ns** |  **2.14 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** |  **4,277.8 ns** | **40.66 ns** | **38.03 ns** |     **?** |       **?** | **1.0529** | **0.0076** |     **-** |   **8.61 KB** |
|           ResolveT |           1 |  2,061.6 ns |  5.59 ns |  5.23 ns |     ? |       ? | 0.6638 | 0.0038 |     - |   5.44 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** |  **7,668.8 ns** | **40.53 ns** | **37.91 ns** |     **?** |       **?** | **1.7700** | **0.0153** |     **-** |  **14.55 KB** |
|           ResolveT |           2 |  3,415.2 ns | 11.21 ns |  9.94 ns |     ? |       ? | 1.0033 | 0.0076 |     - |    8.2 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **10,939.3 ns** | **40.91 ns** | **38.27 ns** |     **?** |       **?** | **2.5024** | **0.0305** |     **-** |  **20.48 KB** |
|           ResolveT |           3 |  4,900.4 ns | 33.25 ns | 31.10 ns |     ? |       ? | 1.3428 | 0.0076 |     - |  10.97 KB |

```Autofac.Benchmarks.Decorators.KeylessNestedSharedInstanceBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |        Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |------------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |    **589.7 ns** |  **2.31 ns** |  **2.16 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** |  **5,151.3 ns** | **22.32 ns** | **20.88 ns** |     **?** |       **?** | **1.2817** | **0.0153** |     **-** |  **10.53 KB** |
|           ResolveT |           1 |  2,567.3 ns |  9.59 ns |  8.97 ns |     ? |       ? | 0.7820 | 0.0076 |     - |    6.4 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** |  **8,085.4 ns** | **37.02 ns** | **34.63 ns** |     **?** |       **?** | **1.9073** | **0.0153** |     **-** |  **15.69 KB** |
|           ResolveT |           2 |  3,610.0 ns | 11.71 ns | 10.95 ns |     ? |       ? | 1.0719 | 0.0076 |     - |   8.77 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **10,403.9 ns** | **45.78 ns** | **42.82 ns** |     **?** |       **?** | **2.5482** | **0.0305** |     **-** |  **20.84 KB** |
|           ResolveT |           3 |  4,702.6 ns | 26.72 ns | 24.99 ns |     ? |       ? | 1.3580 | 0.0153 |     - |  11.15 KB |

```Autofac.Benchmarks.Decorators.KeylessNestedSharedInstanceLambdaBenchmark-report-github.md```
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|             Method | repetitions |        Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |------------ |------------:|---------:|---------:|------:|--------:|-------:|-------:|------:|----------:|
|           **Baseline** |           **?** |    **597.9 ns** |  **1.52 ns** |  **1.42 ns** |  **1.00** |    **0.00** | **0.3271** | **0.0029** |     **-** |   **2.67 KB** |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **1** |  **5,235.9 ns** | **33.76 ns** | **29.93 ns** |     **?** |       **?** | **1.1444** | **0.0153** |     **-** |   **9.38 KB** |
|           ResolveT |           1 |  2,476.8 ns | 18.19 ns | 16.13 ns |     ? |       ? | 0.7095 | 0.0076 |     - |   5.82 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **2** |  **7,684.9 ns** | **38.67 ns** | **36.17 ns** |     **?** |       **?** | **1.7700** | **0.0153** |     **-** |  **14.53 KB** |
|           ResolveT |           2 |  3,554.5 ns | 16.48 ns | 14.61 ns |     ? |       ? | 1.0033 | 0.0114 |     - |    8.2 KB |
|                    |             |             |          |          |       |         |        |        |       |           |
| **ResolveEnumerableT** |           **3** | **10,393.6 ns** | **57.27 ns** | **53.57 ns** |     **?** |       **?** | **2.3956** | **0.0305** |     **-** |  **19.69 KB** |
|           ResolveT |           3 |  4,536.3 ns | 12.50 ns | 11.08 ns |     ? |       ? | 1.2894 | 0.0153 |     - |  10.57 KB |

…